New function in CListModeData and unsigned long for total number of events. [x].Number of list-mode entries now is unsigned long. [x].New function get_total_number_of_events() in the CListModeData. Warning in anything than ROOT files throws an error. [x].If total number of events to store is 1 then LM_TO_PROJDATA will ask the list-mode file for the total number of events. I would prefer to use -1 but now we have unsigned numbers.

+++ src/IO/InputStreamFromROOTFileForCylindricalPET.cxx
@@ -167,7 +167,7 @@ post_processing()
     stream_ptr->SetBranchAddress("rsectorID1",&rsectorID1);
     stream_ptr->SetBranchAddress("rsectorID2",&rsectorID2);
 
-    nentries = static_cast<long long int>(stream_ptr->GetEntries());
+    nentries = static_cast<unsigned long int>(stream_ptr->GetEntries());
     if (nentries == 0)
         error("The total number of entries in the ROOT file is zero. Abort.");
